resolved critical issues
