+++ plugins/vibing/SECURITY_NOTES.md
@@ -1,0 +1,93 @@
+# Security Best Practices - Vibing Plugin
+
+## ✅ Implemented Security Measures
+
+### 1. API Key Handling
+- ✅ API keys loaded from environment variables (priority 1)
+- ✅ Fallback to .env file for local development (priority 2)
+- ✅ Config file support (priority 3)
+- ✅ `.env` files are in `.gitignore` to prevent accidental commits
+
+### 2. Secrets in Logs
+- ✅ **VERIFIED**: No API keys, secrets, or passwords are logged
+- ✅ Error messages do not expose sensitive credentials
+- ✅ Only error types and messages are logged, not actual secret values
+
+### 3. HMAC Signature Implementation
+- ✅ HMAC SHA256 signatures generated correctly
+- ✅ Parameters sorted before signature generation (as per Aster API spec)
+- ✅ Timestamp included in signed requests
+- ✅ Signature verification tested in unit tests
+
+### 4. Error Handling
+- ✅ Structured error responses
+- ✅ No sensitive data in error messages
+- ✅ Graceful handling of API failures
+
+## ⚠️ Recommendations for Future Enhancement
+
+### Rate Limiting
+**Status**: Not currently implemented
+
+**Recommendation**: Add rate limiting to prevent API quota exhaustion:
+
+```python
+# Suggested implementation in aster_client.py
+import time
+from collections import deque
+
+class RateLimiter:
+    def __init__(self, max_calls=1200, period=60):
+        self.max_calls = max_calls
+        self.period = period
+        self.calls = deque()
+    
+    def wait_if_needed(self):
+        now = time.time()
+        # Remove old calls outside the period
+        while self.calls and self.calls[0] < now - self.period:
+            self.calls.popleft()
+        
+        if len(self.calls) >= self.max_calls:
+            sleep_time = self.period - (now - self.calls[0])
+            if sleep_time > 0:
+                time.sleep(sleep_time)
+        
+        self.calls.append(time.time())
+```
+
+**Aster API Limits**:
+- Weight-based rate limiting
+- 1200 requests per minute per IP
+- Order endpoints have higher weight
+
+### Additional Security Recommendations
+
+1. **Request Timeout**: Add timeout to all API requests
+   ```python
+   response = requests.get(url, headers=headers, timeout=10)
+   ```
+
+2. **SSL Verification**: Ensure SSL certificate verification is enabled (default in requests)
+
+3. **Input Validation**: Validate all user inputs before API calls
+
+4. **Idempotency**: Use unique order IDs to prevent duplicate trades
+
+5. **Audit Logging**: Log all trading actions (without secrets) for audit trail
+
+## Testing
+
+Run security tests:
+```bash
+cd animus/smcp/plugins/vibing
+python -m pytest tests/ -v
+```
+
+## Compliance
+
+- ✅ No secrets in version control
+- ✅ No secrets in logs
+- ✅ Proper error handling
+- ⚠️ Rate limiting recommended for production use
+
