+++ plugins/x402/__init__.py
@@ -1,0 +1,10 @@
+"""
+x402 Plugin - SMCP Integration for x402 Vending Machine
+
+Enables agents to interact with the x402 vending machine server.
+"""
+
+__version__ = "1.0.0"
+__author__ = "Animus Team"
+__description__ = "x402 payment-required vending machine integration for SMCP"
+
