+++ plugins/x402/requirements.txt
@@ -1,0 +1,2 @@
+requests>=2.31.0
+
